// src/types/interfaces/ILoginMessages.ts
interface ILoginMessages {
  en: {
    email_required_msg: string;
    email_mismatch_msg: string;
    password_mismatch_msg: string;
    password_empty_msg: string;
    password_length_msg: string;
  };
  ru: {
    email_required_msg: string;
    email_mismatch_msg: string;
    password_mismatch_msg: string;
    password_empty_msg: string;
    password_length_msg: string;
  };
}

export default ILoginMessages;